+++ src/mina-signer/src/signature.unit-test.ts
@@ -18,7 +18,7 @@ import { PrivateKey, PublicKey } from '../../provable/curve-bigint.js';
 import { PrivateKey as PrivateKeySnarky } from '../../lib/signature.js';
 import { p } from '../../snarkyjs-bindings/crypto/finite_field.js';
 import { AccountUpdate } from '../../snarkyjs-bindings/mina-transaction/gen/transaction-bigint.js';
-import { HashInput } from '../../snarkyjs-bindings/provable/provable-bigint.js';
+import { HashInput } from '../../snarkyjs-bindings/lib/provable-bigint.js';
 
 await isReady;
